requirements.txt
Django==5.0.2
django-ninja==1.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
celery==5.3.6
redis==5.0.1
Pillow==10.2.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
drf-yasg==1.21.7
django-cors-headers==4.3.1
django-storages==1.14.2
boto3==1.34.11
django-filter==23.5
django-redis==5.4.0
django-celery-beat==2.5.0
django-celery-results==2.5.1
sentry-sdk==1.39.1 